// src/app/modules/academics/student-information-system/model/students-section-assign-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * StudentSectionAssignView
 */
export interface StudentSectionAssignView { 
    /**
     * Gets or Sets Ids
     */
    ids?: Array<string>;
    /**
     * Gets or Sets ClassId
     */
    classId?: number;
    /**
     * Gets or Sets SectionId
     */
    sectionId?: number;
    /**
     * Gets or Sets SchoolBoardId
     */
    schoolBoardId?: number;
}